Rework date parsing to handle new types
With the introduction of error-chain library I've run into several issues with Vec here. Instead of having the type <Vec<Result<NaiveDate, ParseError>>> which implemented necessary traits for using .collect(), I now have a type of <Vec<Result<NaiveDate>> which requires a different apprroach.

Besides, the way I was trying to parse the date was not that good. There is no need to try every single format if one of them has matched already.